requirements.txt
amqp==5.2.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
billiard==4.2.0
boto3==1.34.66
botocore==1.34.66
celery==5.3.6
certifi==2024.2.2
cffi==1.16.0
cfgv==3.4.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
coverage==7.4.4
cryptography==42.0.5
distlib==0.3.8
Django==4.0
django-cockroachdb==4.1.1
django-cors-headers==3.13.0
django-rest-passwordreset==1.3.0
django-smtp-ssl==1.0
django-storages==1.13.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
drf-spectacular==0.27.0
et-xmlfile==1.1.0
factory-boy==3.3.0
Faker==24.3.0
filelock==3.13.1
gunicorn==20.1.0
identify==2.5.35
idna==3.6
inflection==0.5.1
jmespath==1.0.1
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
kombu==5.3.4
ldap3==2.9.1
msal==1.28.0
nodeenv==1.8.0
numpy
openpyxl==3.1.2
pandas
# pendulum==3.0.0
pendulum==2.1.2
django-celery-beat
platformdirs==4.2.0
pre-commit
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
pyasn1==0.5.1
pycparser==2.21
PyJWT==2.8.0
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.1
referencing==0.32.1
requests==2.31.0
rpds-py==0.16.2
s3transfer==0.10.1
six==1.16.0
sqlparse==0.4.4
time-machine==2.14.0
typing_extensions==4.9.0
tzdata==2023.4
uritemplate==4.1.1
urllib3==2.2.1
vine==5.1.0
virtualenv==20.25.1
wcwidth==0.2.13
whitenoise==6.6.0
django-extensions
django-querycount
django-mailbox==4.10.0
azure-core==1.27.0
azure-identity==1.13.0
azure-storage-blob==12.16.0
imageio-ffmpeg
pillow
moviepy
